+++ src/main/java/net/silentchaos512/mechanisms/block/dryingrack/DryingRackBlock.java
@@ -1,0 +1,95 @@
+package net.silentchaos512.mechanisms.block.dryingrack;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.HorizontalBlock;
+import net.minecraft.block.SoundType;
+import net.minecraft.block.material.Material;
+import net.minecraft.block.material.MaterialColor;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.BlockItemUseContext;
+import net.minecraft.state.DirectionProperty;
+import net.minecraft.state.StateContainer;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.Direction;
+import net.minecraft.util.Hand;
+import net.minecraft.util.Mirror;
+import net.minecraft.util.Rotation;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.BlockRayTraceResult;
+import net.minecraft.util.math.shapes.ISelectionContext;
+import net.minecraft.util.math.shapes.VoxelShape;
+import net.minecraft.world.IBlockReader;
+import net.minecraft.world.World;
+
+import javax.annotation.Nullable;
+
+public class DryingRackBlock extends HorizontalBlock {
+    public static final DirectionProperty FACING = HorizontalBlock.HORIZONTAL_FACING;
+    private static final VoxelShape SHAPE_NORTH = Block.makeCuboidShape(0, 12, 12, 16, 16, 16);
+    private static final VoxelShape SHAPE_SOUTH = Block.makeCuboidShape(0, 12, 0, 16, 16, 4);
+    private static final VoxelShape SHAPE_WEST = Block.makeCuboidShape(12, 12, 0, 16, 16, 16);
+    private static final VoxelShape SHAPE_EAST = Block.makeCuboidShape(0, 12, 0, 4, 16, 16);
+
+    public DryingRackBlock() {
+        super(Properties.create(Material.WOOD, MaterialColor.WOOD).hardnessAndResistance(2f, 3f).sound(SoundType.WOOD));
+        setDefaultState(getDefaultState().with(FACING, Direction.NORTH));
+    }
+
+    @Override
+    public boolean hasTileEntity(BlockState state) {
+        return true;
+    }
+
+    @Nullable
+    @Override
+    public TileEntity createTileEntity(BlockState state, IBlockReader world) {
+        return new DryingRackTileEntity();
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    public boolean onBlockActivated(BlockState state, World worldIn, BlockPos pos, PlayerEntity player, Hand handIn, BlockRayTraceResult hit) {
+        TileEntity tileEntity = worldIn.getTileEntity(pos);
+        if (tileEntity instanceof DryingRackTileEntity) {
+            return ((DryingRackTileEntity) tileEntity).interact(player);
+        }
+        return super.onBlockActivated(state, worldIn, pos, player, handIn, hit);
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    public VoxelShape getShape(BlockState state, IBlockReader worldIn, BlockPos pos, ISelectionContext context) {
+        Direction facing = state.get(FACING);
+        switch (facing) {
+            case NORTH: return SHAPE_NORTH;
+            case SOUTH: return SHAPE_SOUTH;
+            case WEST: return SHAPE_WEST;
+            case EAST: return SHAPE_EAST;
+            default: return SHAPE_NORTH;
+        }
+    }
+
+    @Nullable
+    @Override
+    public BlockState getStateForPlacement(BlockItemUseContext context) {
+        return getDefaultState().with(FACING, context.getPlacementHorizontalFacing().getOpposite());
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    public BlockState rotate(BlockState state, Rotation rot) {
+        return state.with(FACING, rot.rotate(state.get(FACING)));
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    public BlockState mirror(BlockState state, Mirror mirrorIn) {
+        return state.rotate(mirrorIn.toRotation(state.get(FACING)));
+    }
+
+    @Override
+    protected void fillStateContainer(StateContainer.Builder<Block, BlockState> builder) {
+        builder.add(FACING);
+    }
+}