+++ src/main/java/net/silentchaos512/mechanisms/init/ModTileEntities.java
@@ -4,12 +4,17 @@ import net.minecraft.block.Block;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.tileentity.TileEntityType;
 import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.fml.client.registry.ClientRegistry;
+import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.silentchaos512.mechanisms.SilentMechanisms;
 import net.silentchaos512.mechanisms.block.alloysmelter.AlloySmelterTileEntity;
 import net.silentchaos512.mechanisms.block.batterybox.BatteryBoxTileEntity;
 import net.silentchaos512.mechanisms.block.compressor.CompressorTileEntity;
 import net.silentchaos512.mechanisms.block.crusher.CrusherTileEntity;
+import net.silentchaos512.mechanisms.block.dryingrack.DryingRackBlock;
+import net.silentchaos512.mechanisms.block.dryingrack.DryingRackTileEntity;
+import net.silentchaos512.mechanisms.block.dryingrack.DryingRackTileEntityRenderer;
 import net.silentchaos512.mechanisms.block.electricfurnace.ElectricFurnaceTileEntity;
 import net.silentchaos512.mechanisms.block.generator.CoalGeneratorTileEntity;
 import net.silentchaos512.mechanisms.block.wire.WireTileEntity;
@@ -22,6 +27,7 @@ public class ModTileEntities {
     public static TileEntityType<CoalGeneratorTileEntity> coalGenerator;
     public static TileEntityType<CompressorTileEntity> compressor;
     public static TileEntityType<CrusherTileEntity> crusher;
+    public static TileEntityType<DryingRackTileEntity> dryingRack;
     public static TileEntityType<ElectricFurnaceTileEntity> electricFurnace;
     public static TileEntityType<WireTileEntity> wire;
 
@@ -31,6 +37,7 @@ public class ModTileEntities {
         coalGenerator = register("coal_generator", CoalGeneratorTileEntity::new, ModBlocks.coalGenerator);
         compressor = register("compressor", CompressorTileEntity::new, ModBlocks.compressor);
         crusher = register("crusher", CrusherTileEntity::new, ModBlocks.crusher);
+        dryingRack = register("drying_rack", DryingRackTileEntity::new, ModBlocks.DRYING_RACKS.toArray(new DryingRackBlock[0]));
         electricFurnace = register("electric_furnace", ElectricFurnaceTileEntity::new, ModBlocks.electricFurnace);
         wire = register("wire", WireTileEntity::new, ModBlocks.wire);
     }
@@ -40,5 +47,9 @@ public class ModTileEntities {
         type.setRegistryName(SilentMechanisms.getId(name));
         ForgeRegistries.TILE_ENTITIES.register(type);
         return type;
+    }
+
+    public static void registerRenderers(FMLClientSetupEvent event) {
+        ClientRegistry.bindTileEntitySpecialRenderer(DryingRackTileEntity.class, new DryingRackTileEntityRenderer());
     }
 }