+++ src/main/java/net/silentchaos512/mechanisms/item/CraftingItems.java
@@ -1,10 +1,14 @@
 package net.silentchaos512.mechanisms.item;
 
+import net.minecraft.item.Food;
+import net.minecraft.item.Foods;
 import net.minecraft.item.Item;
 import net.minecraft.util.IItemProvider;
 import net.silentchaos512.mechanisms.SilentMechanisms;
 import net.silentchaos512.utils.Lazy;
 import net.silentchaos512.utils.TextUtils;
+
+import javax.annotation.Nullable;
 
 public enum CraftingItems implements IItemProvider {
     REDSTONE_ALLOY_INGOT,
@@ -31,12 +35,33 @@ public enum CraftingItems implements IItemProvider {
     LEAD_INGOT,
     NICKEL_CHUNKS,
     NICKEL_DUST,
-    NICKEL_INGOT;
+    NICKEL_INGOT,
+    ZOMBIE_LEATHER,
+    BEEF_JERKY(Foods.COOKED_BEEF),
+    PORK_JERKY(Foods.COOKED_PORKCHOP),
+    CHICKEN_JERKY(Foods.COOKED_CHICKEN),
+    MUTTON_JERKY(Foods.COOKED_MUTTON),
+    RABBIT_JERKY(Foods.COOKED_RABBIT),
+    COD_JERKY(Foods.COOKED_COD),
+    SALMON_JERKY(Foods.SALMON),
+    ;
 
     private final Lazy<Item> item;
 
     CraftingItems() {
-        this.item = Lazy.of(() -> new Item(new Item.Properties().group(SilentMechanisms.ITEM_GROUP)));
+        this(null);
+    }
+
+    CraftingItems(@Nullable Food food) {
+        this.item = Lazy.of(() -> new Item(createProperties(food)));
+    }
+
+    private static Item.Properties createProperties(@Nullable Food food) {
+        Item.Properties result = new Item.Properties().group(SilentMechanisms.ITEM_GROUP);
+        if (food != null) {
+            result.food(food);
+        }
+        return result;
     }
 
     public String getName() {